Inline reducer types
Similar to previous commits, I'm moving the separate type signature file
in with the implementation. The change is quite bulky because
`createReducer(...)` is a heavy type. I'll try to split this out further
among `handleAction(...)` because as it stands, the file got mean size.
