+++ src/reducers/create-reducer.ts
@@ -1,15 +1,26 @@
-import { Action } from '../types/actions';
-import { CreateReducer } from '../types/create-reducer';
+import { Draft, nothing } from 'immer';
+
 import handleAction from './handle-action';
 import mapActionsToReducers from './map-actions-to-reducers';
 import callOnce from '../utils/call-once';
 import assert from '../utils/assert';
 import { isActionFailure, isOptimisticAction } from '../utils/action-variant';
+import {
+  Action,
+  ActionSuccess,
+  ActionFailure,
+  OptimisticAction,
+  ActionConstant,
+} from '../types/actions';
+import ReducerType from '../constants/reducer-type';
 
-function createReducer<State>(
+export default function createReducer<
+  State,
+  ActionHandlerFactory extends ReducerDefinitionFactory<State>
+>(
   initialState: State,
-  reducerFactory: (...args: any) => any,
-) {
+  reducerFactory: ActionHandlerFactory,
+): (state: undefined | State, action: Action<any>) => State {
   assert(
     typeof reducerFactory === 'function',
     'createReducer(...) expects a function.',
@@ -50,4 +61,107 @@ function createReducer<State>(
   return reducer;
 }
 
-export default createReducer as CreateReducer;
+interface ReducerDefinitionFactory<State> {
+  /**
+   * Use this callback to define all your reducers.
+   * @param handleAction Associates an action with a reducer.
+   * @return A list of reducers defined using `handleAction(...)`.
+   * @example
+   * createReducer(0, handleAction => [
+   *   handleAction(action, reducer),
+   * ])
+   */
+  (handleAction: HandleAction<State>): Array<ReducerDefinition>;
+}
+
+export interface ReducerDefinition {
+  readonly reducerType: ReducerType;
+  readonly actionType: ActionConstant;
+  readonly reducer: (...args: any) => any;
+}
+
+interface HandleAction<State> {
+  /**
+   * Associates a reducer with an action creator.
+   * @param actionCreator An action creator defined with `createAction(...)`.
+   * @param reducer A handler invoked whenever the action is dispatched. It's
+   * okay to mutate state here.
+   * @example
+   * handleAction(increment, state => {
+   *   state.count++
+   * })
+   */
+  <
+    ActionCreator extends (...args: any) => any,
+    Reducer extends (
+      state: Draft<State>,
+      action: SuccessPayload<ActionCreator>,
+    ) => NextState<State>
+  >(
+    actionCreator: ActionCreator,
+    reducer: Reducer,
+  ): ReducerDefinition;
+
+  /**
+   * Associates an action with an error handler.
+   * @param actionCreator An action creator defined with `createAction(...)`.
+   * @param reducer A handler invoked whenever the action is dispatched. It's
+   * okay to mutate state here.
+   * @example
+   * handleAction.error(loadTheme, state => {
+   *   state.theme = THEME_FALLBACK
+   * })
+   */
+  error<
+    ActionCreator extends (...args: any) => any,
+    Reducer extends (state: Draft<State>, error: unknown) => NextState<State>
+  >(
+    actionCreator: ActionCreator,
+    reducer: Reducer,
+  ): ReducerDefinition;
+
+  /**
+   * Associates an async action with an optimistic handler.
+   * @param actionCreator An action creator defined with `createAction.async(...)`.
+   * @param reducer A handler invoked whenever the action is dispatched. It's
+   * okay to mutate state here.
+   * @example
+   * handleAction.optimistic(search, state => {
+   *   state.loading = true
+   * })
+   */
+  optimistic<
+    ActionCreator extends (...args: any) => AsyncGenerator<any, any>,
+    Reducer extends (
+      state: Draft<State>,
+      action: OptimisticPayload<ActionCreator>,
+    ) => NextState<State>
+  >(
+    actionCreator: ActionCreator,
+    reducer: Reducer,
+  ): ReducerDefinition;
+}
+
+type NextState<State> = void | State | typeof nothing;
+
+type ReduxAction<ActionCreator> = ActionCreator extends (
+  ...args: any
+) => infer Payload
+  ? Payload
+  : never;
+
+type SuccessPayload<ActionCreator extends (...args: any) => any> = ReduxAction<
+  ActionCreator
+> extends ActionFailure<any> | ActionSuccess<infer Payload>
+  ? Payload
+  : ReduxAction<ActionCreator> extends ActionSuccess<infer Payload>
+  ? Payload
+  : never;
+
+type OptimisticPayload<
+  ActionCreator extends (...args: any) => AsyncGenerator<any>
+> = ActionCreator extends (
+  ...args: any
+) => AsyncGenerator<OptimisticAction<infer Payload>>
+  ? Payload
+  : never;
